testing permissions and roles through controller

+++ database/migrations/2022_05_30_062706_create_users_permissions_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             // $table->id();
             // $table->timestamps();
 
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('permission_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('permission_id');
 
             //FOREIGN KEY CONSTRAINTS
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
